fix: address PR 1188 review feedback (C1/C2)
- C1: SecretSourceGateway を fail-closed 化。configured_secret_values を
  Result 化し、設定シークレット取得失敗を unwrap_or_default で黙殺せず
  validate_output へ伝播。マスキング漏れによる秘匿情報露出を防止。
- C2: usecase テストの adaptor 層直接依存を解消。WorkflowState の
  wire/view serde 検証4件を presenter/protocol のテストへ移送し、
  usecase 層は domain/usecase 型の検証に限定。
