+++ src-tauri/src/adaptor/gateway/workflow/secret_source_gateway.rs
@@ -1,7 +1,7 @@
 use std::sync::Arc;
 
 use crate::domain::app_config::ConfigSecretRepository;
-use crate::domain::workflow::{secret_masker, SecretSourceGateway};
+use crate::domain::workflow::{secret_masker, SecretSourceGateway, WorkflowError};
 
 #[derive(Clone)]
 pub(crate) struct WorkflowSecretSourceConfigGateway {
@@ -15,12 +15,15 @@ impl WorkflowSecretSourceConfigGateway {
 }
 
 impl SecretSourceGateway for WorkflowSecretSourceConfigGateway {
-    fn configured_secret_values(&self) -> Vec<String> {
-        let mut values = self.config.configured_secret_values().unwrap_or_default();
+    fn configured_secret_values(&self) -> Result<Vec<String>, WorkflowError> {
+        let mut values = self
+            .config
+            .configured_secret_values()
+            .map_err(|e| WorkflowError::External(e.to_string()))?;
         values.extend(secret_masker::collect_secret_values_from_env_vars(
             std::env::vars(),
         ));
-        secret_masker::normalize_secret_values(values)
+        Ok(secret_masker::normalize_secret_values(values))
     }
 }
 
@@ -29,8 +32,8 @@ pub(crate) struct EmptySecretSourceGateway;
 
 #[cfg(test)]
 impl SecretSourceGateway for EmptySecretSourceGateway {
-    fn configured_secret_values(&self) -> Vec<String> {
-        Vec::new()
+    fn configured_secret_values(&self) -> Result<Vec<String>, WorkflowError> {
+        Ok(Vec::new())
     }
 }
 
@@ -50,7 +53,9 @@ mod tests {
         let app_config: Arc<dyn ConfigSecretRepository> =
             Arc::new(AppConfig::new(config, tmp.path().join("config.toml")));
 
-        let secrets = WorkflowSecretSourceConfigGateway::new(app_config).configured_secret_values();
+        let secrets = WorkflowSecretSourceConfigGateway::new(app_config)
+            .configured_secret_values()
+            .unwrap();
 
         assert!(secrets.contains(&"token-12345678".to_string()));
         assert!(secrets.contains(&"https://hooks.example/secret-abcdef".to_string()));
