+++ src-tauri/src/usecase/agent_session/session/mod.rs
@@ -735,11 +735,6 @@ mod tests {
     use super::*;
     use std::collections::{BTreeMap, BTreeSet};
 
-    use crate::domain::workflow::{
-        NodeDefinition, NodeType, StepHistoryEntry, TokenUsage, WorkflowDefinition,
-        WorkflowExecutionState, WorkflowStateSnapshot,
-    };
-
     #[derive(Default)]
     struct TestBackendResolver {
         default_id: Option<String>,
@@ -1595,184 +1590,6 @@ mod tests {
         assert_eq!(content, "hello");
         assert_eq!(thinking, None);
         assert_eq!(activities, None);
-    }
-
-    // ---- WorkflowState wire view serde ----
-
-    fn make_session_test_node(
-        name: &str,
-        node_type: NodeType,
-        instruction: &str,
-    ) -> NodeDefinition {
-        NodeDefinition {
-            name: name.to_string(),
-            node_type,
-            instruction: Some(instruction.to_string()),
-            ..NodeDefinition::default()
-        }
-    }
-
-    fn make_test_workflow_for_session() -> WorkflowDefinition {
-        WorkflowDefinition {
-            variables: Default::default(),
-            name: "review-cycle".to_string(),
-            description: "Test".to_string(),
-            builtin: false,
-            nodes: vec![
-                make_session_test_node("plan", NodeType::Agent, "plan"),
-                make_session_test_node("implement", NodeType::Agent, "implement"),
-                make_session_test_node("review", NodeType::Agent, "review"),
-                NodeDefinition {
-                    name: "report".to_string(),
-                    node_type: NodeType::Approval,
-                    instruction: Some("report".to_string()),
-                    ..NodeDefinition::default()
-                },
-            ],
-        }
-    }
-
-    #[test]
-    fn workflow_state_view_serde_roundtrip() {
-        let state = WorkflowStateSnapshot {
-            execution_id: "exec-1".to_string(),
-            workflow_name: "review-cycle".to_string(),
-            state: WorkflowExecutionState::Running,
-            current_step_index: 2,
-            current_step_name: "review".to_string(),
-            current_session_id: Some("sess-current".to_string()),
-            total_steps: 4,
-            step_history: vec![
-                StepHistoryEntry {
-                    step_name: "plan".to_string(),
-                    completed_at: 1000.0,
-                    result: None,
-                    session_id: None,
-                    token_usage: None,
-                    structured_output: None,
-
-                    run_index: 0,
-                    child_outputs: None,
-                    state: crate::domain::workflow::value_objects::default_step_entry_state(),
-                },
-                StepHistoryEntry {
-                    step_name: "implement".to_string(),
-                    completed_at: 1001.0,
-                    result: Some("done".to_string()),
-                    session_id: Some("sess-1".to_string()),
-                    token_usage: Some(TokenUsage {
-                        input_tokens: 100,
-                        output_tokens: 50,
-                    }),
-                    structured_output: None,
-
-                    run_index: 0,
-                    child_outputs: None,
-                    state: crate::domain::workflow::value_objects::default_step_entry_state(),
-                },
-            ],
-            step_execution_counts: std::collections::HashMap::new(),
-            step_outputs: std::collections::HashMap::new(),
-            workflow_definition: make_test_workflow_for_session(),
-            total_token_usage: TokenUsage {
-                input_tokens: 100,
-                output_tokens: 50,
-            },
-            step_states: std::collections::HashMap::new(),
-            active_parallel_steps: vec![],
-            workflow_variables: std::collections::HashMap::new(),
-            approval_operations: None,
-            started_at: 999.0,
-            updated_at: 1001.0,
-        };
-        let view = crate::adaptor::presenter::workflow::workflow_state_to_view(state);
-        let json = serde_json::to_string(&view).unwrap();
-        let back: crate::adaptor::protocol::workflow::WorkflowStateFieldsView =
-            serde_json::from_str(&json).unwrap();
-        assert_eq!(back.execution_id, "exec-1");
-        assert_eq!(back.workflow_name, "review-cycle");
-        assert_eq!(
-            back.state,
-            crate::adaptor::protocol::workflow::WorkflowExecutionStateView::Running
-        );
-        assert_eq!(back.current_step_index, 2);
-        assert_eq!(back.current_step_name, "review");
-        assert_eq!(back.total_steps, 4);
-        assert_eq!(back.step_history.len(), 2);
-        assert_eq!(back.step_history[0].step_name, "plan");
-        assert_eq!(back.step_history[1].result, Some("done".to_string()));
-    }
-
-    #[test]
-    fn workflow_execution_state_view_failed_tagged_enum_format() {
-        let state = crate::adaptor::protocol::workflow::WorkflowExecutionStateView::Failed {
-            reason: "exit code 1".to_string(),
-        };
-        let json = serde_json::to_string(&state).unwrap();
-        let v: serde_json::Value = serde_json::from_str(&json).unwrap();
-        assert_eq!(v["type"], "failed");
-        assert_eq!(v["reason"], "exit code 1");
-        let back: crate::adaptor::protocol::workflow::WorkflowExecutionStateView =
-            serde_json::from_str(&json).unwrap();
-        assert_eq!(back, state);
-    }
-
-    #[test]
-    fn workflow_execution_state_view_all_variants_serde() {
-        use crate::adaptor::protocol::workflow::WorkflowExecutionStateView;
-
-        let variants = vec![
-            WorkflowExecutionStateView::Running,
-            WorkflowExecutionStateView::WaitingApproval,
-            WorkflowExecutionStateView::Completed,
-            WorkflowExecutionStateView::Failed {
-                reason: "err".to_string(),
-            },
-            WorkflowExecutionStateView::Aborted,
-        ];
-        for state in variants {
-            let json = serde_json::to_string(&state).unwrap();
-            let back: WorkflowExecutionStateView = serde_json::from_str(&json).unwrap();
-            assert_eq!(back, state);
-        }
-    }
-
-    // 撤去済み: `chat_session_with_workflow_state_roundtrip` は ChatSession.workflow_state
-    // フィールド廃止により役目を終えた。WorkflowState の serde roundtrip は
-    // adaptor/protocol/workflow.rs 側の view 型で担保される。在庫 JSON の workflow_state は
-    // serde の unknown_field 既定挙動で silently 読み捨てられる（破棄前提）。
-
-    /// [02] schema 境界: 旧表現（`workflowDefinition.steps`）を含む WorkflowState JSON は
-    /// 新 `Workflow` schema（`nodes` 必須 + `deny_unknown_fields`）として deserialize に失敗する。
-    /// これにより旧表現の進行中状態は新バージョンに引き継がれない。
-    #[test]
-    fn legacy_workflow_state_with_steps_fails_to_deserialize() {
-        let json = r#"{
-            "executionId": "exec-1",
-            "workflowName": "legacy",
-            "state": { "type": "running" },
-            "currentStepIndex": 0,
-            "currentStepName": "x",
-            "totalSteps": 1,
-            "stepHistory": [],
-            "stepExecutionCounts": {},
-            "workflowDefinition": {
-                "name": "legacy",
-                "description": "",
-                "builtin": false,
-                "steps": [{"name":"x","mode":"auto","instruction":"x"}]
-            },
-            "totalTokenUsage": { "inputTokens": 0, "outputTokens": 0 },
-            "stepStates": {},
-            "startedAt": 1.0,
-            "updatedAt": 1.0
-        }"#;
-        let result: Result<crate::adaptor::protocol::workflow::WorkflowStateFieldsView, _> =
-            serde_json::from_str(json);
-        assert!(
-            result.is_err(),
-            "旧 workflowDefinition.steps を含む WorkflowState は新 schema で deserialize 失敗する"
-        );
     }
 
     /// parent ChatSession 廃止後の在庫 JSON 互換性: 旧 `ChatSession.workflowState` フィールドを
